requirements.txt
streamlit
requests
pandas
numpy
PyPDF2
python-docx
rich
langchain
langchain-google-genai
langchain-core
fastapi
uvicorn
pydantic
python-dotenv